Fixed Mintable for echidna task3

+++ src/week6-7-static-analysis/echidna/mintable.sol
@@ -4,16 +4,16 @@ pragma solidity ^0.8.0;
 import "./token.sol";
 
 contract MintableToken is Token {
-    int256 public totalMinted;
-    int256 public totalMintable;
+    uint256 public totalMinted;
+    uint256 public totalMintable;
 
-    constructor(int256 totalMintable_) {
+    constructor(uint256 totalMintable_) {
         totalMintable = totalMintable_;
     }
 
     function mint(uint256 value) public onlyOwner {
-        require(int256(value) + totalMinted < totalMintable);
-        totalMinted += int256(value);
+        require(value + totalMinted < totalMintable);
+        totalMinted += value;
 
         balances[msg.sender] += value;
     }
